+++ src/pyscipopt/__init__.py
@@ -1,9 +1,9 @@
-__version__ = '3.0.2'
+__version__ = '3.0.4'
 
 # required for Python 3.8 on Windows
 import os
 if hasattr(os, 'add_dll_directory'):
-    os.add_dll_directory(os.path.join(os.getenv('SCIPOPTDIR'), 'bin'))
+    os.add_dll_directory(os.path.join(os.getenv('SCIPOPTDIR').strip('"'), 'bin'))
 
 # export user-relevant objects:
 from pyscipopt.Multidict import multidict