+++ src/pyscipopt/scip.pyx
@@ -1076,8 +1076,12 @@ cdef class Model:
         return SCIPgetNLPIterations(self._scip)
 
     def getNNodes(self):
-        """Retrieve the total number of processed nodes."""
+        """gets number of processed nodes in current run, including the focus node."""
         return SCIPgetNNodes(self._scip)
+
+    def getNTotalNodes(self):
+        """gets number of processed nodes in all runs, including the focus node."""
+        return SCIPgetNTotalNodes(self._scip)
 
     def getNFeasibleLeaves(self):
         """Retrieve number of leaf nodes processed with feasible relaxation solution."""
@@ -1363,11 +1367,12 @@ cdef class Model:
         if not onlyroot:
             self.setIntParam("propagating/maxrounds", 0)
 
-    def writeProblem(self, filename='model.cip', trans=False):
+    def writeProblem(self, filename='model.cip', trans=False, genericnames=False):
         """Write current model/problem to a file.
 
         :param filename: the name of the file to be used (Default value = 'model.cip')
         :param trans: indicates whether the transformed problem is written to file (Default value = False)
+        :param genericnames: indicates whether the problem should be written with generic variable and constraint names (Default value = False)
 
         """
         fn = str_conversion(filename)
@@ -1377,9 +1382,9 @@ cdef class Model:
         fn = fn + ext
         ext = ext[1:]
         if trans:
-            PY_SCIP_CALL(SCIPwriteTransProblem(self._scip, fn, ext, False))
+            PY_SCIP_CALL(SCIPwriteTransProblem(self._scip, fn, ext, genericnames))
         else:
-            PY_SCIP_CALL(SCIPwriteOrigProblem(self._scip, fn, ext, False))
+            PY_SCIP_CALL(SCIPwriteOrigProblem(self._scip, fn, ext, genericnames))
         print('wrote problem to file ' + str(fn))
 
     # Variable Functions
@@ -1448,7 +1453,8 @@ cdef class Model:
 
         """
         cdef SCIP_VAR* _tvar
-        PY_SCIP_CALL(SCIPtransformVar(self._scip, var.scip_var, &_tvar))
+        PY_SCIP_CALL(SCIPgetTransformedVar(self._scip, var.scip_var, &_tvar))
+
         return Variable.create(_tvar)
 
     def addVarLocks(self, Variable var, nlocksdown, nlocksup):
@@ -1679,6 +1685,14 @@ cdef class Model:
     def getNConss(self):
         """Retrieve the number of constraints."""
         return SCIPgetNConss(self._scip)
+
+    def getNIntVars(self):
+        """gets number of integer active problem variables"""
+        return SCIPgetNIntVars(self._scip)
+
+    def getNBinVars(self):
+        """gets number of binary active problem variables"""
+        return SCIPgetNBinVars(self._scip)
 
     def updateNodeLowerbound(self, Node node, lb):
         """if given value is larger than the node's lower bound (in transformed problem),
@@ -4070,6 +4084,13 @@ cdef class Model:
 
         """
         PY_SCIP_CALL(SCIPfreeSol(self._scip, &solution.sol))
+
+    def getNSols(self):
+        """gets number of feasible primal solutions stored in the solution storage in case the problem is transformed;
+           in case the problem stage is SCIP_STAGE_PROBLEM, the number of solution in the original solution candidate
+           storage is returned
+         """
+        return SCIPgetNSols(self._scip)
 
     def getSols(self):
         """Retrieve list of all feasible primal solutions stored in the solution storage."""